Separated database in another class
